Added more colors!
Effectively doubled the amount of color palettes available by making it a 50% chance for a pokemon to be shiny. I also removed the stars that come out when you send it out, as that would get annoying pretty quick. There are now always stars on the status screen because every pokemon is special.